requirements.txt
accelerate==1.11.0
annotated-doc==0.0.3
annotated-types==0.7.0
anyio==4.11.0
async-timeout==5.0.1
brotlicffi @ file:///C:/miniconda3/conda-bld/brotlicffi_1762950448512/work
certifi @ file:///C:/miniconda3/conda-bld/certifi_1763135347537/work/certifi
cffi @ file:///C:/miniconda3/conda-bld/cffi_1761832794697/work
charset-normalizer @ file:///C:/miniconda3/conda-bld/charset-normalizer_1761744944645/work
click==8.3.0
colorama==0.4.6
diffusers @ git+https://github.com/huggingface/diffusers.git@3579fdabf905fbcb0952466088573a381f676a55
exceptiongroup==1.3.0
fastapi==0.121.1
filelock @ file:///C:/miniconda3/conda-bld/filelock_1762420019633/work
fsspec==2025.10.0
gmpy2 @ file:///C:/b/abs_d8ki0o0h97/croot/gmpy2_1738085498525/work
h11==0.16.0
hf-xet==1.2.0
hf_transfer==0.1.9
httpcore==1.0.9
httptools==0.7.1
httpx==0.28.1
huggingface-hub==0.36.0
idna @ file:///C:/miniconda3/conda-bld/idna_1761912044311/work
importlib_metadata==8.7.0
Jinja2 @ file:///C:/b/abs_920kup4e6u/croot/jinja2_1741711580669/work
MarkupSafe @ file:///C:/b/abs_a0ma7ge0jc/croot/markupsafe_1738584052792/work
mkl-service==2.4.0
mkl_fft @ file:///C:/Users/dev-admin/mkl/mkl_fft_1730823082242/work
mkl_random @ file:///C:/Users/dev-admin/mkl/mkl_random_1730822522280/work
mpmath @ file:///C:/b/abs_7833jrbiox/croot/mpmath_1690848321154/work
networkx @ file:///C:/b/abs_b054htfn9t/croot/networkx_1737043671910/work
numpy @ file:///C:/b/abs_0123vcxhf8/croot/numpy_and_numpy_base_1725470331966/work/dist/numpy-2.0.1-cp310-cp310-win_amd64.whl#sha256=e824ec3b279f4c5207736c83ac2f2dc79b489e08501a4ea9e3fc178c45db289b
packaging==25.0
peft==0.18.0
pillow @ file:///C:/b/abs_b50vowcrzo/croot/pillow_1738010273782/work
psutil==7.1.3
pycparser @ file:///C:/miniconda3/conda-bld/pycparser_1757496059965/work
pydantic==2.12.4
pydantic-settings==2.12.0
pydantic_core==2.41.5
PySocks @ file:///C:/miniconda3/conda-bld/pysocks_1761753030270/work
python-dotenv==1.2.1
PyYAML @ file:///C:/b/abs_14xkfs39bx/croot/pyyaml_1728657968772/work
redis==7.0.1
regex==2025.11.3
requests @ file:///C:/miniconda3/conda-bld/requests_1762359612241/work
safetensors==0.6.2
shellingham==1.5.4
sniffio==1.3.1
starlette==0.49.3
sympy==1.13.1
tokenizers==0.22.1
torch==2.5.1
torchaudio==2.5.1
torchvision==0.20.1
tqdm==4.67.1
transformers==4.57.1
typer-slim==0.20.0
typing-inspection==0.4.2
typing_extensions @ file:///C:/b/abs_ecq8gc0vbm/croot/typing_extensions_1756281142218/work
urllib3 @ file:///C:/b/abs_393g1ayee_/croot/urllib3_1750775481226/work
uvicorn==0.38.0
watchfiles==1.1.1
websockets==15.0.1
win_inet_pton @ file:///C:/miniconda3/conda-bld/win_inet_pton_1761746276406/work
zipp==3.23.0
